Added admin page. Added roles. Added users management.

+++ src/main/java/edu/karazin/shop/model/BookList.java
@@ -24,7 +24,7 @@ public class BookList {
     private String publisher;
 
     @Column(name = "price")
-    private Integer price;
+    private Double price;
 
     @ManyToMany(cascade = CascadeType.MERGE)
     @LazyCollection(LazyCollectionOption.FALSE)
@@ -52,7 +52,7 @@ public class BookList {
 
     public BookList(){}
 
-    public BookList(String name, Integer pages, List<Author> authors, List<Genre> genres, String publisher, Integer price) {
+    public BookList(String name, Integer pages, List<Author> authors, List<Genre> genres, String publisher, Double price) {
         this.name = name;
         this.pages = pages;
         this.publisher = publisher;
@@ -61,7 +61,7 @@ public class BookList {
         this.genres = genres;
     }
 
-    public BookList(Long id, String name, Integer pages, List<Author> authors, List<Genre> genres, String publisher, Integer price) {
+    public BookList(Long id, String name, Integer pages, List<Author> authors, List<Genre> genres, String publisher, Double price) {
         this.id = id;
         this.name = name;
         this.pages = pages;
@@ -103,11 +103,11 @@ public class BookList {
         this.publisher = publisher;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
